requirements.txt
aliyun-python-sdk-core==2.13.4
apiclient==1.0.4
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-cors-headers==3.0.2
django-earthdistance==1.1.0
django-rest-framework==0.1.0
django-user-agents==0.4.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.3
idna==2.8
jmespath==0.9.4
oauth2client==4.1.3
oauthlib==3.0.1
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
ua-parser==0.8.0
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.25.3
user-agents==2.0
django-filter==2.2.0
whitenoise==2.0.6
gunicorn
django-heroku
setuptools==41.0.0